Minor changes to docs, tests, for PR 614
