+++ test/dummy/db/migrate/20110208155312_set_up_test_tables.rb
@@ -282,7 +282,6 @@ class SetUpTestTables < ActiveRecord::Migration
     remove_index :version_associations, :column => [:version_id]
     remove_index :version_associations, :name => 'index_version_associations_on_foreign_key'
     drop_table :version_associations
-    drop_table :filter_modifier
     drop_table :callback_modifiers
   end
 end
